Add 'description' command to pldgithub.py

+++ bin/pldgithub.py
@@ -5,9 +5,10 @@ import os
 import json
 import requests
 
-if len(sys.argv) < 3 or sys.argv[1] not in ('create', 'delete'):
+if len(sys.argv) < 3 or sys.argv[1] not in ('create', 'delete', 'description'):
     print("""Usage: pldgithub.py create REPO [, REPO2 [, REPO3...]]
-   or: pldgithub.py delete REPO [, REPO2 [, REPO3...]]""")
+   or: pldgithub.py delete REPO [, REPO2 [, REPO3...]]
+   or: pldgithub.py description REPO 'New description'""")
     sys.exit(1)
 
 logpass = tuple(open(os.path.expanduser('~/auth'), 'r').readline().strip().split(':'))
@@ -18,10 +19,15 @@ if sys.argv[1] == 'create':
                 data=json.dumps({'name': newrepo, 'has_issues': False, 'has_wiki': False, 'has_downloads': False}))
         if not req.status_code == 201:
             raise SystemExit("Cannot create repository {} on github".format(newrepo))
-else:
+elif sys.argv[1] == 'delete':
     for cannedrepo in sys.argv[2:]:
         req = requests.delete("https://api.github.com/repos/pld-linux/"+cannedrepo, auth=logpass)
         if not req.status_code == 204:
             raise SystemExit("Cannot delete repository {} from github".format(cannedrepo))
-
+else:
+    (repo, newdesc) = sys.argv[2:4]
+    req = requests.patch("https://api.github.com/repos/pld-linux/"+repo, auth=logpass,
+            data=json.dumps({'name': repo, 'description': newdesc}))
+    if not req.status_code == 200:
+        raise SystemExit("Cannot change description for repository {} on github".format(repo))
